Call "trash" with "-F" argument
Apparently this is necessary to record enough information to allow
Finder to restore the file in its original location.  The official
documentation doesn't actually say so, just that "-F" causes the
trashing to be done by Finder instead of by using the system API,
which isn't the same as saying what effect that has on the end
result.

>  -F  Ask Finder to move the files to the trash, instead of
>      using the system API.

Closes #2.

Co-authored-by: Jonas Bernoulli <jonas@bernoul.li>
  The commit message.