Add new osx-trash-command to deal with incompatible implementation
Several implementations of the "trash" command exist, which are
not compatible.  To deal with this go back to using the bundled
Applescript by default.  Users who notice any slowdown can customize
the new option as appropriate for the "trash" implementation that
they use (or even some other command).  We could continue to first
try "trash", but that would pose the question whether we should do
so with or without the "-F" argument.  It seems better to just use
something that works for everyone, even if that approach is slow.
Lacking a maintainer who actually uses macOS, this seems like the
safest option.

See #2 for backstory.